1. Add new function for updating excel list status base on local pics.

+++ PictureDownloader.py
@@ -3,17 +3,18 @@ import urllib.request
 import os
 import requests
 import sys
+import time
 
 from bs4 import BeautifulSoup
-from time import ctime
 from xlrd import open_workbook
 from xlutils.copy import copy
 
 workDir = os.getcwd()+'/'
 destPath = workDir + 'Download/'
 newExcel = workDir+'PictureUrls_copy.xls'
-srcName = 'PictureUrls.xls'
-maxRetry = 5
+srcExcel = 'PictureUrlsFinal.xls'
+maxRetry = 4
+dosum = False
 
 
 def reName(n, new):
@@ -41,11 +42,13 @@ def readFile(file, rb):
     else:
         sheet = rb.sheet_by_index(0)
         names = sheet.col_values(2)
+        status = sheet.col_values(7)
         subName = sheet.col_values(3)
-        for i in range(1, len(subName)):
+        for i in range(0, len(subName)):
             name = names[i].rstrip().replace(' ', '_')
             result.append({
                 'name': name+'-'+subName[i],
+                'state': status[i],
                 'url': 'https://vhzc.hpb.gov.sg/vhz/phs/capi/cloud/getUserFoodImage.ashx?Service=hpbphs&name='+subName[i]+'&mime=image/jpeg'
             })
             # reName(name+'.jpeg', name+'-'+subName[i])
@@ -62,42 +65,92 @@ def checkFile(name):
     return False
 
 
-def tryDownload(meal, url, cnt=1):
-    if os.path.exists(destPath+meal) == True:
-        # print('*** '+meal+' is exist already.')
+def tryDownload(meal, cnt=1):
+    name = meal['name']
+    if meal['state'] != '':
+        # Update: if last time is successful, just skip this onw
+        return 7
+    if os.path.exists(destPath+name) == True:
+        print('*** '+name+' is exist already.')
         return 0
-    while cnt <= maxRetry:
+    while cnt < maxRetry:
         try:
-            # line ending chars throws 404 issue
-            urllib.request.urlretrieve(url, destPath+meal)
-            print(cnt, '| ' + meal)
+            urllib.request.urlretrieve(meal['url'], destPath+name)
+            print(cnt, '| ' + name)
             break
-        except Exception as e:
+        except:
             cnt += 1
     return cnt
 
 
+def containSub(subs, sub):
+    # for i in range(len(subs)):
+    for i in range(10000):
+        if sub == subs[i]:
+            return i
+    return -1
+
+
+def localFile2Excel():
+    rb = open_workbook(workDir+srcExcel)
+    wb = copy(rb)
+    sh = wb.get_sheet(0)
+    sheet = rb.sheet_by_index(0)
+    subs = sheet.col_values(3)
+    status = sheet.col_values(6)
+
+    res = os.listdir(destPath)
+    for name in res:
+        arr = name.split('-')
+        index = containSub(subs, arr[len(arr) - 1])
+        if index != -1:
+            sh.write(index, 7, 1)
+            if status[index].startswith('All') == True:
+                sh.write(index, 6, 'Succeed at: 8')
+            if status[index].endswith('0') == True:
+                sh.write(index, 6, 'Succeed at: 8')
+        else:
+            sh.write(index, 6, 'All Failed')
+    wb.save(newExcel)
+    return
+
+
 if __name__ == "__main__":
-    rb = open_workbook(workDir+srcName)
+    if dosum == True:
+        localFile2Excel()
+        exit()
+
+    rb = open_workbook(workDir+srcExcel)
     workBook = copy(rb)
     newSheet = workBook.get_sheet(0)
 
-    print('\n>>> start ' + ctime() + '\n')
+    print('\n>>> start ' + time.ctime() + '\n')
     if os.path.exists(destPath) != True:
         distutils.dir_util.mkpath(destPath)
     if os.path.exists(newExcel):
         os.remove(newExcel)
 
-    rows = readFile(srcName, rb)
+    rows = readFile(srcExcel, rb)
     length = len(rows)
-    for i in range(1, length):
-        name = rows[i]['name']
-        url = rows[i]['url']
-        value = tryDownload(name, url)
-        if value != 6 and value != 0:
+    for i in range(10000, 20001):
+        value = tryDownload(rows[i])
+        prin = False
+        if value != maxRetry:
+            prin = True
+            # Update: add new flag to indicate sucessful or not, easy for summation
+            newSheet.write(i, 7, 1)
+        else:
+            prin = i % 10 == 0
+
+        if prin == True:
             print('>>> ' + str(round(i/length*100, 2)) + '%', i, value)
-        res = ('All Failed: ' if value == 6 else 'Succeed at: ') + str(value)
-        newSheet.write(i+1, 6, res)
+
+        if value != 7:
+            res = 'All Failed' if value == maxRetry else (
+                'Succeed at: ' + str(value))
+            newSheet.write(i, 6, res)
+
         if i % 100 == 1:
             workBook.save(newExcel)
+            time.sleep(5)
     workBook.save(newExcel)
